fetch ticket data
